+++ doc/rtd/content/append1/bf/howto.bf.020.rst
@@ -1,0 +1,11 @@
+.. _Howto BF 020:
+Howto BF-020: Accessing Data from OpenML
+========================================
+
+.. automodule:: mlpro.bfexamples.howto_bf_020_accessing_data_from_openml
+
+
+Executable code
+---------------
+.. literalinclude:: ../../../../../src/mlpro/bf/examples/howto_bf_020_accessing_data_from_openml.py
+	:language: python
