+++ doc/rtd/content/append1/bf/howto.bf.021.rst
@@ -1,0 +1,11 @@
+.. _Howto BF 021:
+Howto BF-021: Accessing Data from River
+=======================================
+
+.. automodule:: mlpro.bf.examples.howto_bf_021_accessing_data_from_river
+
+
+Executable code
+---------------
+.. literalinclude:: ../../../../../src/mlpro/bf/examples/howto_bf_021_accessing_data_from_river.py
+	:language: python
