+++ doc/rtd/content/append1/bf/howto.bf.022.rst
@@ -1,0 +1,11 @@
+.. _Howto BF 022:
+Howto BF-022: Accessing Data from Scikit-Learn
+==============================================
+
+.. automodule:: mlpro.bf.examples.howto_bf_022_accessing_data_from_scikitlearn
+
+
+Executable code
+---------------
+.. literalinclude:: ../../../../../src/mlpro/bf/examples/howto_bf_022_accessing_data_from_scikitlearn.py
+	:language: python
