--- doc/rtd/content/append1/oa/howto.oa.001.rst
@@ -1,11 +1,0 @@
-.. _Howto OA 001:
-Howto OA-001: Accessing Data from OpenML
-========================================
-
-.. automodule:: mlpro.oa.examples.howto_oa_001_accessing_data_from_openml
-
-
-Executable code
----------------
-.. literalinclude:: ../../../../../src/mlpro/oa/examples/howto_oa_001_accessing_data_from_openml.py
-	:language: python
