--- doc/rtd/content/append1/oa/howto.oa.002.rst
@@ -1,11 +1,0 @@
-.. _Howto OA 002:
-Howto OA-002: Accessing Data from River
-=======================================
-
-.. automodule:: mlpro.oa.examples.howto_oa_002_accessing_data_from_river
-
-
-Executable code
----------------
-.. literalinclude:: ../../../../../src/mlpro/oa/examples/howto_oa_002_accessing_data_from_river.py
-	:language: python
