--- doc/rtd/content/append1/oa/howto.oa.003.rst
@@ -1,11 +1,0 @@
-.. _Howto OA 003:
-Howto OA-003: Accessing Data from Scikit-Learn
-==============================================
-
-.. automodule:: mlpro.oa.examples.howto_oa_003_accessing_data_from_scikitlearn
-
-
-Executable code
----------------
-.. literalinclude:: ../../../../../src/mlpro/oa/examples/howto_oa_003_accessing_data_from_scikitlearn.py
-	:language: python
