+++ doc/rtd/content/append2/core/images/bf/.$MLPro-BF-Math_class_diagram.drawio.bkp
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2022-09-11T15:43:15.502Z" agent="5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.2.8 Chrome/102.0.5005.167 Electron/19.0.15 Safari/537.36" etag="lrWJyf0PSCKbIbGGw-tw" version="20.2.8" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7V1pc+I4E/41qZ19q5LygQ18JEAyk8lNkp3MF5fBApQY2/EBIb/+leQD2xKOIcYQ8M7WDJYPSf10t1qtbulIbE/ez23VGl+ZGtCPBE57PxI7R4IgNCQZ/YNL5n6J1BTrfsnIhppfxi8KevADBIVcUOpBDTiJB13T1F1oJQsHpmGAgZsoU23bnCUfG5p6slZLHQGqoDdQdbr0P6i546BUlmqLGz8BHI2DqgVRDLo8UcOng644Y1UzZ7EisXsktm3TdP1fk/c20DH5QsL892v+n375Kp9f3Dlv6uPp74frp2P/Y2ervBL1wQaGu/anP7i76YV8qpxqU+9uCMTe24d3LARoOu48pBjQEAGDS9N2x+bINFS9uyg9tU3P0AD+LIeuFs9cmqaFCnlU+AJcdx5wg+q5JioauxM9uIu6Yc//4PdPpPDyOX6v8x583L+aR1daC/MEuuzr5uDVLzqDevjhnIQKCOqYnj0AGdQJgHdVewTcjOcCImLKxbgugOEcmBOAeoEesIGuunCaZE014PBR9NwCRPQjwHEFTINWT1XdC2rqodanYXZmcKKrBsZzaBpuiLiIrlUdjgz0e4CoCGxUMAW2C5FMtYIbLsb5dDCGunapzk0P08Zx1cFreHU6Nm34gT6rhsig27YbMIQgJ57o4TcDhG3goGduQwD5qOhSddzgmYGp66rlwD5pMH5kggCCxqnpuuYk/FCCRyPR9Ztim6+RNsAlQ8RBbVM3bUIaUVNBYziInozdkQcN0B+iOyNb1SBqY+zekPyH3wooe5b8anR/NSbFtAfvmWwVamsxQD7Q1ceizDX8ktlC9Qly8NQ4pvVEjlvOjEGF90g9q8YIEX1pjXyNo+rjWfUJcrI6VUeMZqguOMWwOZQIRJ1dXypkSioUDU4UHTq0bCCCuzE50MHQXSoFjqUOoDG6JM90aouS+6C7uMhE7w51woFjqGnAIBzqqq7qMzHmUcuEhkvoIZ2i/xHV2lg9SqhBbXTNL67R//hxG3GfgVhUhYSPAJKQGcBSwuCwTEXxOYfNk7B9xk5peNdRbd51ozs8c1/a3qX0oN737sUBPBbYIMKAZSoMP8dQEsrDkG1yUBhS2OmQjEsxi4HW2Z8AO0EQEWMlQPKBmCbHPIW2SKMtMpDV1T7Qb00HutDE37f9Z1OIbwvUek4939gQpiItlwo0oKsoP/6tJDMniI0StevFhWI/vjlgBocK+OiKyvTjhmE4qpqGFewPC/99JLY6cAIMB0lAhWpeVHl+2wqXr1G4oulMgCvUKihzQynmVLMbk1Baz2IkDW9C0KyQzI2kXNsykktlEhuzFZL5kWzSU86NITn4+2iJgzZnvv7tdOfD++vbs2PGqImoPTOIbiXTywrM3PNLvkQTSJ78hM+92/Nb66L9h1dOZ+fyOHK6LMAcmNYcYWmAWTTVFDsPtgcqWHPDWivRBGLCSs85VcsChoaAdUAloCs4gEq0gLJc7TEkozkJKv7RH56AKeq0c9LF/1yphjoCNg1w5f3+kvdbVx0n+Pi38ITXpLwGX0PM4Nxv5wl/fYByc3xuPLy072d9TxABf8MY49rKaavXVXrdB+X+UHVhPTeTbcEZzsSRHtRiOF5XOO6gQ5yJIz2pj+H4t8KxYB94ETja3hgKj/3pCzx9aVn/vcw6Xu+YzwSyc1MhmXdOz5Xonfk5ferX74fK+fR1KPK/G52LB44xp28r3afu9YNyevN43Wnd/+r29hROCqa8CGe4TUt00TDb16TgRPP4PcVvA+IobXvZnxZHxVAnQHHQNOpgozdWx7GxbUuHgWNfdQD2wVQo5p125I3p2lwAByOUioijbhqjCse8OIolGjkDr3Nl37ZvJ8NH/fdz9wZ2n96WalVdxV2ocMzrBti2dcNLNJCeASuVmhvCMiNvmKJIu+QIgpUoroSjuP1QG8bQ2MeOXNWG4GBjVFcHssxAmylU/36cSoM7Z3Zz9Wd+ZWiNHmts1IAzsKFF4j8rIHMCWWacDVsiaSQp9A493Hh1WPMG3Wwq3phnDJhhwLFFfDvou6g9iKtEHtPzqC0ctThcttYPK+FvKPrb0RxY7CRX4tqF94BM0cTOP//gjwf3iq5C9bnIr6NYShGzdmPNj5lcm2l+zBSIJLjwSuLjVNSN4j6vm6MRJEx0aY5O2srlzblyffPw89f1Obp/sKE0a+jQnDYOHy5QFu5WpY1VHIMKtcMNP1198VgqcfZ/cXz3dn5uXNzcPr48efX6lSWOGJYqCQmPhqsKzPxg1rcd0sGOCg/tgwrK3FDKXImuAKZc0hZqJJfYCKuwzI+lUKI3gOmeEzKwxOZiBWZ+MKVtL3uwc6iw7V/BmB/GxrbtHrZMLqZwFZj5I+e4MjOPWWCyZXIxXa7AzA+mWKLlwwyeoy0fJwlm2heyv26D4oPp6tK2l0bYhm3M61QJ6woxyyVaQ8zMONp9ECQ8+smOqOjaNMC/R2ILPRVLtaoAzrenB1+incSOSucpsMrY4e8durEN/tDVc/Ax/HuxvR++2NbufnygSePb+zFJGMBV3PZ+5FXUV3UeeyDg5sWXb3HBUZRiVg/3aIu2dxNqjTh7UG/wnPDJK+iH34x1E8zYrg96ibWUTSUz2If7Evswe1mjuScrybZg7qHAFqnd/6RmavnVb2rwXkqtrAY7czmBTs298nTUI2TkoQkZusPchnKsWvgnGiV0P3l2NoYu6KERARfPbNVKwmwBG41CJDM3fOl2UZROZh0Oh8KAmcyqyX1Zwqm4iexYK1hyF9ZSOfkzVHlKMmVG/lL0VGJNndvQSEGbAh003N70X/C+vBRuVQL1nm0f2uTTLFnLm1LH1zN4Mqjw+yRNs80oSjgUDUnHARjCmbpiJ7OlmS2m/YUK+rBaobirudLMFtOOQgq7A4rmKwbUEjcPZTaYXvSOxfL54oljpP5nLQT2EDxMxYCbO7WhtiF06eQU4i5EdDwIP2ExKPJbdyOxY8UigaywXGEX0W2bQqF/JgamP8vDSzHmEDs1xhD/ducWNtcHKt5bqw8IvqaNpkWkUfjQlollGsBYvEieBDqYoMKTI0HWMSX7Nvo18mmaLmmbNnF8kGUg10w6mTkc+YRrxY6DsDXZe2ps35+QYvjE3JHmqYLcCZxMeX8aDH9CkzV5EwpwKEyPtZ8jieu3Wy9CxxMmmqf8YuVetMnOYQHS/poR5qszDQyx2YbKQvQjlkLgOyc7iPNu+I1qMmvSvCm/0c+f9nQ8ce6ceXc0ulJbj+fXaq4Um5FtetYXiRIdEhXo+6P4MUxMYvFpl7lQY0hFrcFSvkVIBZtctCshVIvYk0a6GXKb/OaZvt4UOfJfvIhzwtO2wjI+fjulcRN4hE/hCo/9z+D1Pb5mvdOfuAQjYGjkiWX6HBGizyjz+7NM9a9S8hllhsMkZT7rqczqaGb/WH1BRorBrC1oJ67JHvV/kBQKMomJ/fp3eQOC3n7aCyZc6F91gpWf0Xcsnxx00bJuJR77hTtn2ebIBlhp5/hsoUWkeiQlEBmYxBwotQH3QNXm4UjlYTMkopkP+2oMmhTdZUityH+rSUURcpgoSTRedV00lntkoQcrSIFD6nJskmSS4lC6BjPMk15fhwOsnBeVnvnVhYgRaxVOLN8UxSbrilqpSEDXhI9VRSaZY0opqgl9q5lsSuqBYrAhwJxBY4xsKkLuz0Aaq7jvZJqRgAmNwjZYDkuxkpDogkLRtlA0WU1QVpORgPvRpBdZQYRYIT3P4m3fHQGgScqt0+UkX8W6f7UldlrS96IUa97Gsj0usUlPLucLa+KUXP/EtcamQBwnyxxZF01PjgLapCZH9Nxq2Upm9vzi81nSFq162qfSAXiLeTKPpQPmGNEqMdyWLRiEAF6m7kcLFnkXNpKBVWQF359MoH6sDMynoS0BaMe8FDogogkZz0gkkBjISRnAsQNimifNZlOO/tQTFaMZ8gnHiRwvNerob5lLxdv40T9UqAxViVQ7adQafPinLq5Uy5KAHLorqaUXczgMt8w7Wy+YLL3yHshFUWFhbHahF+9++Xody8jQNie7KyXcYUqJKJyIjZrMiXwd/51aIypISLIrWVVG5jEaZ0lMYVxNL19SbFwFSy0Nlho2BoDtaO03pJqEP1hssNRXTYzI455gN4YTtsaQzzRzf+s4KTYl6fVeShwOKEIjW2V8JXqKxV4FRGiwG0yv/lY6bgUd9/li0m7quErJLSElHepeKbllOuPbKLlGpeS+U9R7QUqusuSWkZI+NaNScst0xg4qORf2en8Nc9hvvr0bz7yoCK+3YbRa2Zl5qOl+NmhdCq+f4zcXCaHkKndG6DpppjS0hWeRlpgHuPks0pogZb/xZW9hJl1j6qfrL+tUo/IOj8rFhLlJ9TSX8ozArSVRs/t0gCdzBw56qemQT9ApZAwueVNyxoFk5DeqqGV4E2t+ovp6vYJ09xLT2Ici0UJ5yNZyQbCWmJrGRpVesozlph3y0fGrY1lmJtrFhWI/vjlgBocK+OiKyvTjZkkmmp99oh32/rqro1lqRhpzu5EliYVk26oD3jBwDSjLTEhjq1k2lr5JVEG5ApTy1u0gOhzCqaBcB8r8E94NKVj2pqzkMtofsAI0/6wzXPbfGqAZornY8JEkJ1l+aYVufnQZ/qKSDxyQKbTK9e0X59pnxpxz3NnZl9z3bKoFWi7uvs/a7q849/3XwKaVc5eE8P+4Iv/Qkls5yffNSc5LKSe50GzmtOTX268tSqmI5oDCTjjJ2RKSIy73gBxy2VqkcNd5AQ45dovpmaIGka4xBtiEsfjQfLGEQ7BdCgK1zJPos8bW+M6x/mDWY3lYq5Fs30ay9OLrCsu9e7XzKFM4qljTT1VeEfH05Q5j9B4wsWEMRMMY2MNhLP/Erhisyxzdnlut95fplTu5O/v19ueWa8+n/SgmdQ8m5l87hYFJnYA4n4bPhS6Z8ifgWc2mo9MukRxXJsvemyz19IZ7Qj3v5DvahW8vbBa2yqOkQ8GbQio6Szi++4BGs1imxtjJMDU2istghAHXVCh+jmKZkWnMFgvV/KFwUEuMS2M2OCMsbf8mDJsCscyANGaLabeXqmlEw6JJ4GIpFhccia3Avqzw3ckQtfOPzvhPu6c9PU1Uri6eXXDTR8YkAUdQhENoJaq7GaLGbDI7Qi2S1QrKnQxRY25L3qCwKsEdQ5Mzr9eE2Yfd95q4H4bW/xBG7qX0ys8hr2j129BuqdJPN8kI5YSvkFdXzT6VqDPmOEmOc1WO7NPUG1/OPs0ia0zdn3nGwGUdfVz59vbNtyc21vbtRbtJ7IVvj6nDGU4haFieqzj+FrItf+3+AAyiTM2xk24+JqC0g0hBuqdCdG1Ey3T5MVtcnZJYPKgluvyYUkpbJLFMVLYCbvvb//PE7GwLRy30DW69H9YylbCJOoiDZKBjCkZbfnCHMMMuhlPz+jWlAvyazBbTfs2Jav0gCBNGjTszyTk2yzZ2qRBe4kPJnWxSAMTdFhzf9f5c/H6+Gz45T43RnXjB0kbLMMZkC0T8K5oi1A7xD++hUsgfHVUQK8k5w6PETRkrjFXmBzCxcFY+Kh0iSpCDQVJnYpzC4ytPd+FUtaH/5ER1xwD9hXHGLwbzdcbBVekJ/GEeJShQu8w3GCMH8yjBrACWL2Uo5thycDsnCaaPXRQFibYHRZFBq0hXf4VYwH3sqL/vWqI96tfgsX3bARZDB1Nyo2K5SIkJqpscy+Qw53TfThgy+WilE2M2dmoms4k5dpvdCqvX0pzOyVvndHqESOl+enzoegMdatA/n3GPOT1/ovIGOB1d2iY+BGvhX0RUG1+ZGsBP/B8=</diagram></mxfile>
