+++ doc/rtd/content/append2/core/images/bf/.$MLPro-BF-Math_class_diagram.drawio.dtmp
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2022-09-11T15:47:25.412Z" agent="5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/20.2.8 Chrome/102.0.5005.167 Electron/19.0.15 Safari/537.36" etag="HoLJ8XG54lfztlwmDryc" version="20.2.8" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7V1pc+I4E/41qZ19q5LygQ35SDgyk8lNkp3MF5fBMigxtuMDQn79K8kHNhLGEGNI8M7WDJYPSf10t1qtbulIbI3fzx3VHl1ZGjCOBE57PxLbR4IgNCQZ/YNLZkGJdCrWg5KhA7WgjJ8X9OAHCAu5sNSHGnBTD3qWZXjQThcOLNMEAy9VpjqONU0/pltGulZbHQKqoDdQDbr0P6h5o7BUlmrzGz8BHI7CqgVRDLs8VqOnw664I1WzpokisXMkthzL8oJf4/cWMDD5IsL892v2n3H5Kp9f3Llv6uPZ74frp+PgY911Xon74ADT2/jTH9zd5EI+U860iX+nA7H39uEfCyGarjeLKAY0RMDw0nK8kTW0TNXozEvPHMs3NYA/y6Gr+TOXlmWjQh4VvgDPm4XcoPqehYpG3tgI76JuOLM/+P0TKbp8Tt5rv4cfD65m8ZXWxDyBLvuGNXgNirrQiD6ck1AhQV3LdwYggzoh8J7qDIGX8VxIREy5BNeFMJwDawxQL9ADDjBUD07SrKmGHD6Mn5uDiH6EOK6BadjqiWr4YU091PpFmN0pHBuqifHULdOLEBfRtWrAoYl+DxAVgYMKJsDxIJKpZnjDwzifDUbQ0C7VmeVj2rieOniNrs5GlgM/0GfVCBl02/FChhDk1BM9/GaIsANc9MxtBCAfF12qrhc+M7AMQ7Vd2CcNxo+MEUDQPLM8zxpHH0rxaCy6QVMc6zXWBrhERxzUsgzLIaQRNRU09EH8ZOKOPGiAvo7uDB1Vg6iNiXs6+Q+/FVK2m/5qfH89JsW0B++ZbBVpazFEPtTVx6LMNYKS6Vz1CXL41Cih9USOW86MYYX3SD2r5hARfWmNfI2j6uNZ9QlyujrVQIxmqh44w7C5lAjEnd1cKmRKKhQNjhUDurRsIIJ7CTkwgO4tlQLXVgfQHF6SZ9q1ecl92F1cZKF3dYNw4AhqGjAJh3qqpwZMjHnUtqDpEXpIZ+h/RLUWVo8SalALXfPza/Q/ftxB3GciFlUh4SOAJGQKsJQwOCxTUazmsFkatlXstAjvJqrNv2509K730vIvpQf1vncvDuCxwAYRhixTYbgaQ0koD0O2yUFhSGFnQDIuJSwGWmevAHaMICLGSojkAzFNjnkKbZFGW2Qga6h9YNxaLvSghb/vBM8uIL4rUOs59XxjS5iKtFwq0ISeovz4t5LMnCA2StSuFxeK8/jmginUFfDREZXJxw3DcFQ1DSvYHzb++0hstuEYmC6SgArVvKjy/K4VLl+jcEXTmRBXqFVQ5oZSzKlmtyahtJ7FSJr+mKBZIZkbSbm2YySXyiQ2Zisk8yN5Sk85t4bk4O+jLQ5anPX6t92Z6ffXt91jxqiJqD01iW4l08sKzNzzS75EE0ge/4TPvdvzW/ui9YdXzqbn8ih2uszBHFj2DGFpgmk81RTbD44PKlhzw1or0QRiwkrPOVXbBqaGgHVBJaBrOIBKtICyXO0JJOM5CSr+0ddPwAR12j3p4H+uVFMdAocGuPJ+f8r7baiuG378S3jCa1Jeg68hZnDul/OEvz5A+XR0bj68tO6nfV8QAX/DGONaylmz11F6nQfl/lB1YT03k+3AGc7EkR7UEjheVzjuoUOciSM9qU/g+LfCsWAfeBE4Ov4ICo/9yQs8e2na/71M237vmM8Esn1TIZl3Ts+V6J35OXnq1+915Xzyqov870b74oFjzOlbSuepc/2gnN08Xreb9786vQrO3G7TEl00zCafUnCieXyFX178pF0v+9PiqJjqGCgumkYdbPTG+jg2dm3pMHDsqy7APpgKxbzTjrwxXdsL4GCEUhFxNCxzWOGYF0exRCNn4LevnNvW7Vh/NH4/d25g5+ltqVY1VNyFCse8boBdWze8RAPpm7BSqbkhLDPyhimKtEuOIFiJ4lo4irsPtWEMjX3syFUdCA42RnV9IMsMtJlA9e/HmTS4c6c3V39mV6bW6LHGRg24AwfaJP6zAjInkGXG2bAlkkaSQu/Qw43XhzVv0M224o15xoAZBRzbxLeDvovag7hK5DE9j1rCUZPDZRv9sFP+hqK/Hc+BxXZ6Ja5VeA/IFE1s//MP/nh4r+gq1ICLgjqKpRQxa7fW/ITJtZ3mJ0yBWIILryQ5TsXdKO7zhjUcQsJEl9bwpKVc3pwr1zcPP39dn6P7BxtKs4EOzWnj8NECZeFuVdpYxTGoUDvc8NP1F4+lEmf/F8d3b+fn5sXN7ePLk1+vX9nikGGpkpDweLiqwMwPZn3XIR3sqPDIPqigzA2lzJXoCmDKJW2hxnKJjbAKy/xYCiV6A5juOSEDS2wuVmDmB1Pa9bIHO4cK2/4VjPlhbOza7mHL5HwKV4GZP3KOKzPzmAUmWybn0+UKzPxgiiVaPszgOdrycdNgLvpCKrfBGvBKu14aYRu2Ca9TJaxrxCyXaA0xM+No90GY8BgkO6Kia8sE/x6JTfRUItWqAjjfnh58iXYSOyqdp8AqY4e/d+glNvhDV8/hx/Dv+fZ++GJXu/vxoSZNbu/HJGEIV3Hb+5FXUV/VWeKBkJvnX77FBUdxilk92qMt3t5NqDWS7EG9wXPCilfQj6AZmyaYsV0f9BJrKZtKZrAP9yn2YfayRnNPVpJtwdxDgS1Su/9JpwvLr0FTw/cW1Mp6sDOXE+jU3CvfQD1CRh6akKE7zG0oR6qNf6JRwgiSZ6cj6IEeGhFw8dRR7TTMNnDQKEQyc6OXbudFi8msuq4LA2Yyqyb3ZQmn4qayY+1wyV3YSOXkz1DlKcmUGflL8VOpNXVuSyMFbQq00XB703/B+/JSuFUJ1N9s+9BTfpEla3lT6vh6Bk+GFX6dpGm2GUUJh6Ih6TgAQzhTV+xltjSzxbS/UEEfVisU9zVXmtli2lFIYXdA0XzFgFri5qHMBtOL3olYvkA8cYzU/+y5wB6Ch6kYcHOnNtS2hC6dnELchYiOB+EnLAZFfuduJHasWCyQFZZrpMPv2hSK/DMJMINZHl6KsXTs1BhB/Nub2dhcH6h4b60+IPhaDpoWkUbhQ1vGtmUCc/4ieRIYYIwKT44E2cCU7Dvo1zCg6WJJy3KI44MsA3lW2snM4cgnXCt2HEStyd5TY/f+hAWGT80daZ4qyJ3AyZT3p8HwJ5yyJm9CAQ6FybH2cyhx/VbzRWj7wljzlV+s3IsW2TksRDpYM8J81dWAjs02VBahH7MUAt892UOc98NvVJNZk+Zt+Y1+/nQmo7F75846w+GV2nw8v1ZzpdgMHcu3P0mU+JCoUN8fJY9hYno0KGJJjGTdWoOle4sQCja1aE9CpBWxI430MmI2+c23ArUpcuS/ZBHnRodtRWV88vaCwk3BET2FKzwOPoOX9/ia/U5/4hIMgamRJ5apc0SIPqMs6M8yzb9OySrK6HqaMqt6KrM6mtk/Vl+QjWIyawvbiWtyhv0fJIOCzGESv/5d3oCwtyt7wYQL/auOse4z+64dkIMuWtat1GO/cOdsxxo6AOvsHJ8ttIhUj6QEIvuSWAOlNuAeqNosGqh8bIXENAtgX49B06K7DKk1+W89qShCDlMlqcarnoeGcp+s82AFKXBIXY4skktSHErXYIp50u8bcICV87zSblBdhBgxVuHYDixRbLGuqZWKBHRD+FhVZJI5oZTimtC3TtNNWXigGGwIMF1ojpBJRci9CqSRivtOZhkpmNAo7IDlsBQrCakuKBRtC0WT1QRlPRkJuR/NeZERRIgV0bObbPv+CABNUm6TLqf5KtH9qx2x05K+F6VY8zaW7XBJzHly+V5Y86b08ieuNTED4jhZ5siy6OLcKKTNwtyInlotW8jMnl6sniTt0KqnXSptgHeYJ9NYOl6OEaySwG3ZekEE4OXC/Xi9Iu+6RjquiizgB5MJ1I+1gVkZ2RKCdsxLkf8hmpAJPCOPQGIgJ2UAx46HOT05PT2V4z/1VMVognzCcSLHS406+lvmFsJtguAfKlKGqkSqnTRqDT76UxfXqmVJPA7dlYWVF0vXox3zupvFki0uvIdyUVRUGJtd6LW7X4FexzKiO9Z4f6WEO0wpEYUTsVGTOZGv478XlogKEpLsStaVkVmCxlkSUxhX06uXFBtXsVJLY6X0xgCw/az9hlST8AeLjZX6rIkRO9xT7MbwwdYY8rnI3F86TIpNSXq5lxKHAwrQyFYZnwmeYrFXAQEa7AbTi7+VjltDx61eS9pPHVcpuSWkpCPdKyW3TGd8GSXXqJTcVwp6L0jJVZbcMlLSh2ZUSm6ZzthDJefBXu+vaen907d385kXFeH1NgpWKzsxDzU9SAatS9H1c/LmPB+UXOVOCN0ky5SGtvAk0hLTALefRFoTpOw3Pu0tzKRrQv10gmWdalTe41G5mCg3qb7IpTwjmnFJ0Ox3Or+TuQEHvdR0yAfoFDIGl7wnOeM8MvIbVdQ0/bE9O1EDvV5Bun95aewzkWihPGRruSBYS8xMY6NKL1kmUtMO+eT49bEsMxHt4kJxHt9cMIW6Aj46ojL5uFmSiBYkn2iHvb3u+miWmpDG3G1kSV4h2bXqgPcL3ADKMvPR2GqWjWVgElVQrgGlvHM7iA6HcCsoN4Ey/4R3SwqWvScruYy3B6wAzT/rjJb9dwZohmjO93skyUl2UFqhmx9dhr+o5PMGZAqtcn37xbn2mTHnHNftfsp9z6ZaqOWS7vus3f6Kc99/DmxaOXdICP+PK/IPLbmVk/y7Ocl5acFJLpye5rTkN9uuLU6piOeAwl44ydkSkiMu94AcctlapHDXeQEOOXaL6ZmiBpGuMQfYhLH5yHyxhUOwXQoCtcyD6LPG1uTGscFg1mN5WKuR7LuNZIuLr2ss936rjUeZwlHFmq5UeUXE05c7jNF7wCSGMRAPY+AbDmP5J3bFYF3m6PbcbL6/TK688V3319ufW641m/TjmNRvMDH/3CEMTOqExFkZPhe5ZMqfgGc1m45Ou0RyXJks395kqS/utyfU806+4034voXNwlZ5lHQoeE9IxWAJx1cf0GgWy9QYexmmxkZxGYww5JoKxdUolhmZxmyxUM0fCge1xLg0ZoMzwtK+34RhWyCWGZDGbDHt9lI1jWhYNAmcL8XigiOxGdqXFb57GaJ2/tEe/Wn1tKenscrVxe4FN3lkTBJwBEU0hFaiup8haswmsyPUYlmtoNzLEDXmruQNCqsS3DE0OfN6TZh92H+vifdhav0PYehdSq/8DPKKVr+N7JYq/XSbjFBO+Ap5dd3sU4k6Yo6T5CRX5cg+XXjj09mnWWRNqPuubw481snHlW/vu/n2xMbGvr14N4lv4dtj6nCGUwiatu8pbrCFbDNYuz8AgyhTc+ylm48JKO0gUpDuqRDdGNEyXX7MFleHJBYPaokuP6aU0hZJIhOVrYBbwfb/PDE7W8JRE32D2+yHvUwlbKMO4iAZGJiC8ZYf3CHMsIvh1Lx+TakAvyazxbRfc6zaPwjChFGTzkxyjs2yjV0qhJf4UHInmxQAcacJR3e9Pxe/n+/0J/epMbwTL1jaaBnGmGyhiH9GU0TaIfnhb6gU8kdHFcRKcs7wKHFbxgpjlfkBjG2clY9KdUQJcjDIwpkYZ/D4yjc8OFEdGDw5Vr0RQH9hnPGL4XydcXDV4gT+ME8SFKhd5huMkYN5kmBWAMunMhRzbDm4k4MEqVMXRUGi7UFRZNAq1tWfIRbwHtvq77um6Az7NXjs3LaBzdDBlNyoWC4WxATVTY5lcplzui8nDJl8tNaJMVs7NJPZxBy7ze6E1WuLnM7JO+d0eoRY0P30+NDxBwbUYHA+4zfm9PyJylvgdHTpWPgQrLl/EVFtdGVpAD/xfw==</diagram></mxfile>
