+++ doc/rtd/content/append2/core/mlpro.oa.rst
@@ -1,18 +1,6 @@
 Online Adaptivity
 ================
 
-Stream Providers and Streams
-----------------------------
-
-.. image:: images/oa/MLPro-OA-Streams_class_diagram.drawio.png
-
-.. automodule:: mlpro.oa.models_streams
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-
 Serial Processing
 -----------------
