+++ doc/rtd/content/append2/wrappers/river.rst
@@ -11,4 +11,4 @@ River
 
 Cross References
 ----------------
-    + :ref:`Howto OA-002: Accessing Data from River<Howto OA 002>`
+    + :ref:`Howto BF-021: Accessing Data from River<Howto BF 021>`
