+++ src/mlpro/bf/examples/howto_bf_020_accessing_data_from_openml.py
@@ -1,7 +1,7 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
 ## -- Package : mlpro
-## -- Module  : howto_oa_001_accessing_data_from_openml.py
+## -- Module  : howto_bf_020_accessing_data_from_openml.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
 ## -- yyyy-mm-dd  Ver.      Auth.    Description
