+++ src/mlpro/bf/math.py
@@ -18,17 +18,22 @@
 ## -- 2021-12-03  1.3.2     MRD      Fix Set.append() due to the usage of max() on empty list
 ## -- 2022-01-21  1.4.0     DA       New class TrendAnalyzer
 ## -- 2022-02-25  1.4.1     SY       Class Dimension extended by auto generated ID
+## -- 2022-09-11  1.5.0     DA       - Class Dimension: new method set_boundaries (event)
+## --                                - Class TrendAnalyzer removed
+## --                                - Code reformatting
 ## -------------------------------------------------------------------------------------------------
 
 """
-Ver. 1.4.1 (2022-02-25)
+Ver. 1.5.0 (2022-09-11)
 
-This module provides basic mathematical classes .
+This module provides basic mathematical classes.
 """
 
 import numpy as np
 from itertools import repeat
 import uuid
+from mlpro.bf.various import Log
+from mlpro.bf.events import *
 
 
 
@@ -36,7 +41,7 @@ import uuid
 
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
-class Dimension:
+class Dimension (EventManager):
     """
     Objects of this type specify properties of a dimension of a set.
 
@@ -58,17 +63,32 @@ class Dimension:
         List with minimum and maximum value (optional)
     p_description : str
         Description of dimension (optional)
+    p_logging
+        Log level (see constants of class Log). Default: Log.C_LOG_ALL
 
     """
 
-    C_BASE_SET_R = 'R'  # real numbers
-    C_BASE_SET_N = 'N'  # natural numbers
-    C_BASE_SET_Z = 'Z'  # integer numbers
-    C_BASE_SET_DO = 'DO'  # (big) data objects (like images, point clouds, ...)
+    C_BASE_SET_R        = 'R'   # real numbers
+    C_BASE_SET_N        = 'N'   # natural numbers
+    C_BASE_SET_Z        = 'Z'   # integer numbers
+    C_BASE_SET_DO       = 'DO'  # (big) data objects (like images, point clouds, ...)
 
-    ## -------------------------------------------------------------------------------------------------
-    def __init__(self, p_name_short, p_base_set=C_BASE_SET_R, p_name_long='', p_name_latex='', p_unit='',
-                 p_unit_latex='', p_boundaries=[], p_description='') -> None:
+    C_EVENT_BOUNDARIES  = 0     # raised by method set_boundaries()
+
+## -------------------------------------------------------------------------------------------------
+    def __init__( self, 
+                  p_name_short, 
+                  p_base_set=C_BASE_SET_R, 
+                  p_name_long='', 
+                  p_name_latex='', 
+                  p_unit='',
+                  p_unit_latex='', 
+                  p_boundaries=[], 
+                  p_description='',
+                  p_logging=Log.C_LOG_NOTHING ):
+
+        EventManager.__init__(self, p_logging=p_logging)
+
         self._id = str(uuid.uuid4())
         self._name_short = p_name_short
         self._base_set = p_base_set
@@ -76,46 +96,63 @@ class Dimension:
         self._name_latex = p_name_latex
         self._unit = p_unit
         self._unit_latex = p_unit_latex
-        self._boundaries = p_boundaries
         self._description = p_description
 
-    ## -------------------------------------------------------------------------------------------------
+        self.set_boundaries(p_boundaries=p_boundaries)
+
+
+## -------------------------------------------------------------------------------------------------
     def get_id(self):
         return self._id
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_name_short(self):
         return self._name_short
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_base_set(self):
         return self._base_set
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_name_long(self):
         return self._name_long
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_name_latex(self):
         return self._name_latex
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_unit(self):
         return self._unit
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_unit_latex(self):
         return self._unit_latex
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_boundaries(self):
         return self._boundaries
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
+    def set_boundaries(self, p_boundaries):
+        self._boundaries = p_boundaries
+        self._raise_event( p_event_id=self.C_EVENT_BOUNDARIES, p_event_object=Event(p_raising_object=self, p_boundaries=p_boundaries) )
+
+
+## -------------------------------------------------------------------------------------------------
     def get_description(self):
         return self._description
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def copy(self):
         return self.__class__(p_name_short=self._name_short,
                               p_base_set=self._base_set,
@@ -127,6 +164,9 @@ class Dimension:
                               p_description=self._description)
 
 
+
+
+
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
 class Set:
@@ -134,12 +174,13 @@ class Set:
     Objects of this type describe a (multivariate) set in a mathematical sense.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def __init__(self) -> None:
         self._dim_list = []
         self._dim_ids = []
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def add_dim(self, p_dim: Dimension):
         """
         Raises the dimensionality of the set by adding a new dimension.
@@ -151,7 +192,8 @@ class Set:
         self._dim_ids.append(p_dim.get_id())
         self._dim_list.append(p_dim)
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_dim(self, p_id) -> Dimension:
         """
         Returns the dimension specified by it's unique id.
@@ -159,7 +201,8 @@ class Set:
 
         return self._dim_list[self._dim_ids.index(p_id)]
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_num_dim(self):
         """
         Returns the dimensionality of the set (=number of dimensions of the set).
@@ -167,7 +210,8 @@ class Set:
 
         return len(self._dim_list)
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_dim_ids(self):
         """
         Returns the unique ids of the related dimensions.
@@ -175,7 +219,8 @@ class Set:
 
         return self._dim_ids
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def spawn(self, p_id_list: list):
         """
         Spawns a new class with same type and a subset of dimensions specified
@@ -194,7 +239,8 @@ class Set:
 
         return new_set
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def copy(self, p_new_dim_ids=True):
         new_set = self.__class__()
 
@@ -207,7 +253,8 @@ class Set:
 
         return new_set
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def append(self, p_set, p_new_dim_ids=True):
         if p_new_dim_ids:
             for dim_id in p_set.get_dim_ids():
@@ -217,6 +264,9 @@ class Set:
                 self.add_dim(p_set.get_dim(dim_id))
 
 
+
+
+
 ## -------------------------------------------------------------------------------------------------
 ## -------------------------------------------------------------------------------------------------
 class DataObject:
@@ -224,18 +274,23 @@ class DataObject:
     Container class for (big) data objects of any type with optional additional meta data.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def __init__(self, p_data, *p_meta_data) -> None:
         self._data = p_data
         self._meta_data = p_meta_data
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_data(self):
         return self._data
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_meta_data(self) -> tuple:
         return self._meta_data
+
+
+
 
 
 ## -------------------------------------------------------------------------------------------------
@@ -245,24 +300,28 @@ class Element:
     Element of a (multivariate) set.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def __init__(self, p_set: Set) -> None:
         self._set = p_set
         self._values = list(repeat(0, self._set.get_num_dim()))
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_related_set(self) -> Set:
         return self._set
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_dim_ids(self) -> list:
         return self._set.get_dim_ids()
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_values(self):
         return self._values
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def set_values(self, p_values):
         """
         Overwrites the values of all components of the element.
@@ -273,13 +332,18 @@ class Element:
 
         self._values = p_values
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_value(self, p_dim_id):
         return self._values[self._set.get_dim_ids().index(p_dim_id)]
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def set_value(self, p_dim_id, p_value):
         self._values[self._set.get_dim_ids().index(p_dim_id)] = p_value
+
+
+
 
 
 ## -------------------------------------------------------------------------------------------------
@@ -289,12 +353,13 @@ class ElementList:
     List of Element objects.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def __init__(self, ):
         self._elem_list = []
         self._elem_ids = []
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def add_elem(self, p_id, p_elem: Element):
         """
         Adds an element object under it's id in the internal element list.
@@ -307,13 +372,18 @@ class ElementList:
         self._elem_ids.append(p_id)
         self._elem_list.append(p_elem)
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_elem_ids(self) -> list:
         return self._elem_ids
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def get_elem(self, p_id) -> Element:
         return self._elem_list[self._elem_ids.index(p_id)]
+
+
+
 
 
 ## -------------------------------------------------------------------------------------------------
@@ -324,9 +394,12 @@ class MSpace(Set):
     space.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def distance(self, p_e1: Element, p_e2: Element):
         raise NotImplementedError
+
+
+
 
 
 ## -------------------------------------------------------------------------------------------------
@@ -337,9 +410,12 @@ class ESpace(MSpace):
     implements the Euclidian norm.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def distance(self, p_e1: Element, p_e2: Element):
         return np.sum((np.array(p_e1.get_values()) - np.array(p_e2.get_values())) ** 2) ** 0.5
+
+
+
 
 
 ## -------------------------------------------------------------------------------------------------
@@ -350,7 +426,7 @@ class Function:
     multivariate input space to elements of a multivariate output space.
     """
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def __init__(self, p_input_space: MSpace, p_output_space: MSpace, p_output_elem_cls=Element):
         """
         Parameters:
@@ -363,7 +439,8 @@ class Function:
         self._output_space = p_output_space
         self._output_elem_cls = p_output_elem_cls
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def map(self, p_input: Element) -> Element:
         """
         Maps a multivariate abscissa/input element to a multivariate ordinate/output element. 
@@ -373,95 +450,7 @@ class Function:
         self._map(p_input, output)
         return output
 
-    ## -------------------------------------------------------------------------------------------------
+
+## -------------------------------------------------------------------------------------------------
     def _map(self, p_input: Element, p_output: Element):
         raise NotImplementedError
-
-
-## -------------------------------------------------------------------------------------------------
-## -------------------------------------------------------------------------------------------------
-class TrendAnalyzer:
-    """
-    Template class for trend analysis. Values can be added and the implemented algorithm decides
-    about the trend.
-
-    Parameters
-    ----------
-    p_horizon : int
-        This number of last values is buffered as base for the trend computation.
-
-    """
-
-    C_TREND_POSITIVE = 1
-    C_TREND_NEGATIVE = -1
-    C_TREND_NEUTRAL = 0
-
-    ## -------------------------------------------------------------------------------------------------
-    def __init__(self, p_horizon: int):
-        self._horizon = p_horizon
-        self._trend = self.C_TREND_NEUTRAL
-        self._buffer = []
-
-    ## -------------------------------------------------------------------------------------------------
-    def add_value(self, p_value: float) -> int:
-        """
-        Adds a new value to the buffer and computes the trend of the buffered values by calling the
-        custom methodd _compute_trend().
-
-        Parameters
-        ----------
-        p_value : float
-            Value to be added to the internal buffer.
-
-        Returns
-        -------
-        trend : int
-            Trend of the buffered values. Possible values are C_TREND_POSITIVE, C_TREND_NEGATIVE, C_TREND_NEUTRAL.
-
-        """
-
-        self._buffer.append(p_value)
-        if len(self._buffer) > self._horizon: self._buffer.pop()
-
-        trend = self._compute_trend(np.asarray(self._buffer))
-        if trend > 0:
-            self._trend = self.C_TREND_POSITIVE
-        elif trend < 0:
-            self._trend = self.C_TREND_NEGATIVE
-        else:
-            self._trend = self.C_TREND_NEUTRAL
-
-        return self._trend
-
-    ## -------------------------------------------------------------------------------------------------
-    def _compute_trend(self, p_values: np.ndarray) -> int:
-        """
-        Custom method for trend algorithm.
-
-        Parameters
-        ----------
-        p_values : np.ndarray
-            Numpy array with values to be analyzed.
-
-        Returns
-        -------
-        trend : float
-            Where a value >0 means a positive trend, 0 a neutral trend and <0 a negative trend.
-
-        """
-
-        raise NotImplementedError
-
-    ## -------------------------------------------------------------------------------------------------
-    def get_trend(self) -> int:
-        """
-        Returns the trend of the currently buffered values.
-
-        Returns
-        -------
-        trend : int
-            Valid values are: -1 for negative trend, 0 for a neutral trend and 1 for a positive trend.
-
-        """
-
-        return self._trend
