+++ src/mlpro/bf/streams.py
@@ -1,6 +1,6 @@
 ## -------------------------------------------------------------------------------------------------
 ## -- Project : MLPro - A Synoptic Framework for Standardized Machine Learning Tasks
-## -- Package : mlpro.oa
+## -- Package : mlpro.bf
 ## -- Module  : models_streams.py
 ## -------------------------------------------------------------------------------------------------
 ## -- History :
@@ -27,7 +27,7 @@ Model classes for stream providers and streams.
 from mlpro.bf.various import *
 from mlpro.bf.ml import *
 from mlpro.bf.math import *
-import datetime
+from datetime import datetime
 
 
 
@@ -142,22 +142,22 @@ class Stream (Mode, LoadSave, ScientificObject):
         self._kwargs = p_kwargs.copy()
 
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def get_id(self) -> str:
         return self._id
 
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def get_name(self) -> str:
         return self.C_NAME
 
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def get_url(self) -> str:
         return self.C_SCIREF_URL
 
 
-    ## -------------------------------------------------------------------------------------------------
+## -------------------------------------------------------------------------------------------------
     def get_num_features(self) -> int:
         return self._num_instances
