+++ src/mlpro/wrappers/openml.py
@@ -31,7 +31,7 @@ https://docs.openml.org/APIs/
 import numpy
 from mlpro.bf.various import ScientificObject
 from mlpro.wrappers.models import Wrapper
-from mlpro.oa.models import *
+from mlpro.bf.streams import *
 from mlpro.bf.math import *
 import openml
