+++ src/mlpro/wrappers/sklearn.py
@@ -26,7 +26,7 @@ https://scikit-learn.org
 
 from mlpro.bf.various import ScientificObject
 from mlpro.wrappers.models import Wrapper
-from mlpro.oa.models import *
+from mlpro.bf.streams import *
 from mlpro.bf.math import *
 import sklearn
 from sklearn import datasets
